Adjusts the api-auth to immediately retrieve access and refresh tokens when initalizing a OAuthTokenProvider.
It might occur that the necessary code received from the authentication endpoint is already invalid when accessing the token api due to delays within an application flow or debugging.
Due to that the initialization of the OauthTokenProvider now explicitly retrieves the token on instantiation, no call to trigger the retrieval is required anymore.